refactor: move logic into context provider (#4265)
