+++ packages/shared/src/components/onboarding/OnboardingHeader.tsx
@@ -18,7 +18,7 @@ import { PlusUser } from '../PlusUser';
 import { IconSize } from '../Icon';
 import { TypographyType } from '../typography/Typography';
 import { PlusFreeTrialAlert } from '../plus/PlusFreeTrialAlert';
-import { usePaymentContext } from '../../contexts/PaymentContext';
+import { usePaymentContext } from '../../contexts/payment/context';
 
 type OnboardingHeaderProps = {
   showOnboardingPage: boolean;
