+++ packages/shared/src/components/plus/PlusIOS.tsx
@@ -1,22 +1,16 @@
 import type { ReactElement } from 'react';
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
-import { useQuery } from '@tanstack/react-query';
+import { useQueryClient } from '@tanstack/react-query';
 import { PlusInfo } from './PlusInfo';
-import type {
-  OpenCheckoutFn,
-  ProductOption,
-} from '../../contexts/PaymentContext';
+import type { OpenCheckoutFn } from '../../contexts/payment/context';
+import { usePaymentContext } from '../../contexts/payment/context';
 import type { CommonPlusPageProps } from './common';
-import { isNullOrUndefined, promisifyEventListener } from '../../lib/func';
-import { PlusPriceType, PlusPriceTypeAppsId } from '../../lib/featureValues';
-import { useFeature } from '../GrowthBookProvider';
-import { featureIAPProducts } from '../../lib/featureManagement';
+import { promisifyEventListener } from '../../lib/func';
 import { webappUrl } from '../../lib/constants';
-import { useAuthContext } from '../../contexts/AuthContext';
 import { Button, ButtonVariant } from '../buttons/Button';
-import { sendMessage, WebKitMessageHandlers } from '../../lib/ios';
+import { WebKitMessageHandlers } from '../../lib/ios';
 import { useToastNotification } from '../../hooks';
 import { DEFAULT_ERROR } from '../../graphql/common';
 import Toast from '../notifications/Toast';
@@ -27,105 +21,23 @@ const PlusTrustRefund = dynamic(() =>
 
 const PlusFAQs = dynamic(() => import('./PlusFAQ').then((mod) => mod.PlusFAQ));
 
-type IAPProduct = {
-  attributes: {
-    description: {
-      standard: string;
-    };
-    icuLocale: string;
-    isFamilyShareable: number;
-    kind: string;
-    name: string;
-    offerName: string;
-    offers: {
-      currencyCode: string;
-      discounts: unknown[]; // Specify a more detailed type if known
-      price: string;
-      priceFormatted: string;
-      recurringSubscriptionPeriod: string;
-    }[];
-    subscriptionFamilyId: string;
-    subscriptionFamilyName: string;
-    subscriptionFamilyRank: number;
-  };
-  href: string;
-  id: string;
-  type: string;
-};
-
 export const PlusIOS = ({
   shouldShowPlusHeader,
 }: CommonPlusPageProps): ReactElement => {
   const router = useRouter();
   const { displayToast } = useToastNotification();
-  const { user } = useAuthContext();
-  const productIds = useFeature(featureIAPProducts);
   const [selectedOption, setSelectedOption] = useState<string | null>(null);
+  const { productOptions, openCheckout } = usePaymentContext();
+
+  const queryClient = useQueryClient();
 
   const selectionChange: OpenCheckoutFn = useCallback(({ priceId }) => {
     setSelectedOption(priceId);
   }, []);
 
   const onContinue = useCallback(() => {
-    sendMessage(WebKitMessageHandlers.IAPSubscriptionRequest, {
-      productId: selectedOption,
-      appAccountToken: user?.subscriptionFlags?.appAccountToken,
-    });
-  }, [selectedOption, user?.subscriptionFlags?.appAccountToken]);
-
-  const productList = useMemo(() => Object.keys(productIds), [productIds]);
-
-  const { data: productOptions, ...query } = useQuery({
-    queryKey: ['iap-products'],
-    queryFn: async () => {
-      const products = promisifyEventListener(
-        'iap-products-result',
-        (event) => {
-          const productsRaw = !isNullOrUndefined(event?.detail)
-            ? JSON.parse(event.detail)
-            : [];
-
-          return productsRaw
-            ?.map((product: IAPProduct): ProductOption => {
-              const duration = productIds[
-                product.attributes.offerName
-              ] as PlusPriceType;
-
-              return {
-                label: product.attributes.name,
-                value: product.attributes.offerName,
-                price: {
-                  amount: parseFloat(product.attributes.offers[0].price),
-                  formatted: product.attributes.offers[0].priceFormatted,
-                  monthlyAmount: parseFloat(product.attributes.offers[0].price),
-                  monthlyFormatted: product.attributes.offers[0].priceFormatted,
-                },
-                extraLabel: product.attributes?.description?.standard,
-                appsId:
-                  product.attributes.offerName === 'annualSpecial'
-                    ? PlusPriceTypeAppsId.EarlyAdopter
-                    : PlusPriceTypeAppsId.Default,
-                duration,
-                durationLabel:
-                  duration === PlusPriceType.Yearly ? 'year' : 'month',
-                trialPeriod: null,
-              };
-            })
-            .sort((a: { value: string }, b: { value: string }) => {
-              // Make sure that the products are sorted in the same order as the product list
-              // because the native code does not guarantee the order of the products
-              const aIndex = productList.indexOf(a.value);
-              const bIndex = productList.indexOf(b.value);
-              return aIndex - bIndex;
-            });
-        },
-      );
-
-      sendMessage(WebKitMessageHandlers.IAPProductList, productList);
-
-      return products;
-    },
-  });
+    openCheckout({ priceId: selectedOption });
+  }, [openCheckout, selectedOption]);
 
   useEffect(() => {
     promisifyEventListener('iap-purchase-result', (event) => {
@@ -165,7 +77,9 @@ export const PlusIOS = ({
         <div className="flex gap-2">
           <Button
             onClick={() => {
-              query.refetch();
+              queryClient.refetchQueries({
+                queryKey: ['iap-products'],
+              });
             }}
             variant={ButtonVariant.Float}
           >
