+++ packages/shared/src/components/plus/PlusMobile.tsx
@@ -3,8 +3,8 @@ import React, { useCallback, useState } from 'react';
 import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
 import { PlusInfo } from './PlusInfo';
-import type { OpenCheckoutFn } from '../../contexts/PaymentContext';
-import { usePaymentContext } from '../../contexts/PaymentContext';
+import type { OpenCheckoutFn } from '../../contexts/payment/context';
+import { usePaymentContext } from '../../contexts/payment/context';
 import type { CommonPlusPageProps } from './common';
 import { useGiftUserContext } from './GiftUserContext';
 import { webappUrl } from '../../lib/constants';
