+++ packages/shared/src/contexts/payment/Paddle.tsx
@@ -1,18 +1,12 @@
-import type { ReactElement, ReactNode } from 'react';
+import type { ReactElement } from 'react';
 import React, {
   useCallback,
-  useContext,
   useEffect,
   useMemo,
   useRef,
   useState,
 } from 'react';
-import type {
-  Environments,
-  Paddle,
-  PaddleEventData,
-  TimePeriod,
-} from '@paddle/paddle-js';
+import type { Environments, Paddle, PaddleEventData } from '@paddle/paddle-js';
 import {
   CheckoutEventNames,
   getPaddleInstance,
@@ -20,59 +14,23 @@ import {
 } from '@paddle/paddle-js';
 import { useQuery } from '@tanstack/react-query';
 import { useRouter } from 'next/router';
-import { useAuthContext } from './AuthContext';
-import { plusSuccessUrl } from '../lib/constants';
-import { LogEvent } from '../lib/log';
-import { usePlusSubscription } from '../hooks';
-import { feature } from '../lib/featureManagement';
-import { PlusPriceType, PlusPriceTypeAppsId } from '../lib/featureValues';
-import { getPrice } from '../lib';
-import { useFeature } from '../components/GrowthBookProvider';
-import { checkIsExtension, isIOSNative } from '../lib/func';
-import { usePixelsContext } from './PixelsContext';
-
-export type ProductOption = {
-  label: string;
-  value: string;
-  price: {
-    amount: number;
-    formatted: string;
-    monthlyAmount: number;
-    monthlyFormatted: string;
-  };
-  currencyCode?: string;
-  currencySymbol?: string;
-  extraLabel: string;
-  appsId: PlusPriceTypeAppsId;
-  duration: PlusPriceType;
-  durationLabel: 'month' | 'year';
-  trialPeriod: TimePeriod | null;
-};
-
-interface OpenCheckoutProps {
-  priceId: string;
-  giftToUserId?: string;
-}
-
-export type OpenCheckoutFn = (props: OpenCheckoutProps) => void;
-
-export interface PaymentContextData {
-  openCheckout?: OpenCheckoutFn;
-  paddle?: Paddle | undefined;
-  productOptions?: ProductOption[];
-  earlyAdopterPlanId?: string | null;
-  isPlusAvailable: boolean;
-  giftOneYear?: ProductOption;
-  isPricesPending: boolean;
-  isFreeTrialExperiment: boolean;
-}
-
-const PaymentContext = React.createContext<PaymentContextData>(undefined);
-export default PaymentContext;
-
-export type PaymentContextProviderProps = {
-  children?: ReactNode;
-};
+import { useAuthContext } from '../AuthContext';
+import { plusSuccessUrl } from '../../lib/constants';
+import { LogEvent } from '../../lib/log';
+import { usePlusSubscription } from '../../hooks';
+import { feature } from '../../lib/featureManagement';
+import { PlusPriceType, PlusPriceTypeAppsId } from '../../lib/featureValues';
+import { getPrice } from '../../lib';
+import { useFeature } from '../../components/GrowthBookProvider';
+import { checkIsExtension } from '../../lib/func';
+import { usePixelsContext } from '../PixelsContext';
+import type {
+  PaymentContextProviderProps,
+  OpenCheckoutProps,
+  PaymentContextData,
+  ProductOption,
+} from './context';
+import { PaymentContext } from './context';
 
 export const PaddleSubProvider = ({
   children,
@@ -316,17 +274,3 @@ export const PaddleSubProvider = ({
     </PaymentContext.Provider>
   );
 };
-
-export const PaymentContextProvider = ({
-  children,
-}: PaymentContextProviderProps): ReactElement => {
-  if (isIOSNative()) {
-    // TODO: Implement native payment context
-    return <PaddleSubProvider>{children}</PaddleSubProvider>;
-  }
-
-  return <PaddleSubProvider>{children}</PaddleSubProvider>;
-};
-
-export const usePaymentContext = (): PaymentContextData =>
-  useContext(PaymentContext);
